added: devices context
added devices context and refactored routing
