+++ client/src/components/Loading.js
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import {ReactComponent as LoadingMeters}  from '../assets/loadingMeters.svg';
 import {ReactComponent as LoadingDots}  from '../assets/loadingDots.svg';
+import { motion } from 'framer-motion'
 
-
-let Loading = ({loadedWhen}) => {
+let Loading = ({error}) => {
 	let [ready, setReady] = useState(false)
 
 	setTimeout(()=>{
@@ -11,10 +11,15 @@ let Loading = ({loadedWhen}) => {
 	},10)
 
 	return(
-			<div className="w-full h-screen bg-neutral-100 flex flex-col justify-center
-			 items-center" >
+			<motion.div
+				className="w-full h-screen bg-neutral-100 flex flex-col justify-center items-center"
+				initial={{opacity: 0}}
+				animate={{opacity: 1}}
+				exit={{opacity: 0}}
+			 >
 				<LoadingMeters className={`mx-auto h-1/4 fill-neutral-800 transform transition-all duration-500 ease-in opacity-0 ${ ready ? 'opacity-100' : '' }` } />
-			</div>
+				<p className="py-2 text-red-500">{ error ? 'Connection Error' : ''}</p>
+			</motion.div>
 	)	
 }
