+++ client/src/context/Devices.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const DevicesContext = createContext([])
